+++ control/__init__.py
@@ -1,0 +1,5 @@
+from control.sos import sos_ctrl
+
+__all__ = [
+    "sos_ctrl",
+]