+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Current Changelog
 
-
+0.0.0:
+* Added yaml config loading
+* Added imgui skeleton
+* Added controller code
+* Added sequencer code
+* Added logger code
 
 # Older Changelogs